add document resolver, add proper changes to service.xml, clean up rabbitMqPersistanceHandler
