+++ DocumentResolver/Type/Odt.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Simgroep\ConcurrentSpiderBundle\DocumentResolver\Type;
+
+use VDB\Spider\Resource;
+
+/**
+ * Description of Odt
+ *
+ * @author lkalinka
+ */
+class Odt implements DocumentTypeInterface
+{
+    public function getData(Resource $resource)
+    {
+
+    }
+
+}
