+++ DocumentResolver/Type/Rtf.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Simgroep\ConcurrentSpiderBundle\DocumentResolver\Type;
+
+use VDB\Spider\Resource;
+
+/**
+ * Description of RTF
+ *
+ * @author lkalinka
+ */
+class Rtf implements DocumentTypeInterface
+{
+    public function getData(Resource $resource)
+    {
+
+    }
+
+}
