+++ DocumentResolver/Type/Word2007.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Simgroep\ConcurrentSpiderBundle\DocumentResolver\Type;
+
+use VDB\Spider\Resource;
+
+/**
+ * Description of Word2007
+ *
+ * @author lkalinka
+ */
+class Word2007 implements DocumentTypeInterface
+{
+
+    public function getData(Resource $resource)
+    {
+
+    }
+
+}
